requirements.txt
# LiveKit packages
livekit>=0.11.0
livekit-agents>=0.8.0
livekit-api>=0.6.0

# Google Gemini
google-generativeai>=0.3.0

# RAG and embeddings
faiss-cpu>=1.7.4
sentence-transformers>=2.2.0

# Utilities
python-dotenv>=1.0.0
numpy>=1.24.0

# Web server for token generation
flask>=3.0.0
flask-cors>=4.0.0

# Optional but recommended
aiohttp>=3.9.0